wad_strip: fix case where TEXTUREx lump is before PNAMES
The PNAMES lump must be built before the TEXTUREx lump. The wad_strip tool
keeps the original order of the lumps, so if TEXTUREx was before PNAMES the
TEXTUREx lump would be built with incorrect patch indicies. Fix this by
always building the PNAMES lump first, but still preserving the lump order.